Keep consistency of the bundle digest between the ER & fraud proof generation
This commit also move ExtrinsicDigest to sp-domains, otherwise import it in the
block-preprocessor will cause cyclic dependency
