+++ crates/sp-domains-fraud-proof/src/fraud_proof.rs
@@ -5,15 +5,14 @@ use sp_core::H256;
 use sp_domain_digests::AsPredigest;
 use sp_domains::proof_provider_and_verifier::StorageProofVerifier;
 use sp_domains::{
-    BundleValidity, DomainId, ExecutionReceipt, HeaderHashFor, HeaderHashingFor, InvalidBundleType,
-    SealedBundleHeader,
+    BundleValidity, DomainId, ExecutionReceipt, ExtrinsicDigest, HeaderHashFor, HeaderHashingFor,
+    InvalidBundleType, SealedBundleHeader,
 };
 use sp_runtime::traits::{Block as BlockT, Hash as HashT, Header as HeaderT, NumberFor};
 use sp_runtime::{Digest, DigestItem};
 use sp_std::vec::Vec;
 use sp_trie::StorageProof;
 use subspace_runtime_primitives::{AccountId, Balance};
-use trie_db::TrieLayout;
 
 type ExecutionReceiptFor<DomainHeader, CBlock, Balance> = ExecutionReceipt<
     NumberFor<CBlock>,
@@ -558,33 +557,6 @@ pub struct InvalidDomainBlockHashProof<ReceiptHash> {
     pub bad_receipt_hash: ReceiptHash,
     /// Digests storage proof that is used to derive Domain block hash.
     pub digest_storage_proof: StorageProof,
-}
-
-/// Represents the extrinsic either as full data or hash of the data.
-#[derive(Clone, Debug, Decode, Encode, Eq, PartialEq, TypeInfo)]
-pub enum ExtrinsicDigest {
-    /// Actual extrinsic data that is inlined since it is less than 33 bytes.
-    Data(Vec<u8>),
-    /// Extrinsic Hash.
-    Hash(H256),
-}
-
-impl ExtrinsicDigest {
-    pub fn new<Layout: TrieLayout>(ext: Vec<u8>) -> Self
-    where
-        Layout::Hash: HashT,
-        <Layout::Hash as HashT>::Output: Into<H256>,
-    {
-        if let Some(threshold) = Layout::MAX_INLINE_VALUE {
-            if ext.len() >= threshold as usize {
-                ExtrinsicDigest::Hash(Layout::Hash::hash(&ext).into())
-            } else {
-                ExtrinsicDigest::Data(ext)
-            }
-        } else {
-            ExtrinsicDigest::Data(ext)
-        }
-    }
 }
 
 /// Represents a valid bundle index and all the extrinsics within that bundle.
